Update version number from 0.13 to 0.14.
Fix in  PHP::assign_globals  for an issue with assignments
to the superglobal variables  $_REQUEST, $_SERVER,  and
$_ENV . For some reason, the assignment does not happen
if there has not been some other reference to them. The
current workaround for this issue might not be the best
solution, but it does work.

Added test cases for assignment to superglobals.

Added XS(PHP_spoof_rfc1867) function. The purpose of this
function is to add entries to the internal hashtable that
PHP uses to track uploaded files. If you are manipulating
PHP's  $_FILES  variable outside of PHP, you will need to
do this to get PHP's  is_uploaded_file()  function to return
a true value.  See  main/rfc1867.c  in the PHP source for
more information.

The Perl accessor for this function is  PHP::_spoof_rfc1867.
For now there is no documentation for this method in
PHP.pm .

Added test cases for  PHP::_spoof_rfc1867  and its effect
on PHP's  is_uploaded_file()  function.